Also add configuration test for spotbugs
